+++ lib/input/redis_streams.go
@@ -17,10 +17,6 @@ func init() {
 Pulls messages from Redis (v5.0+) streams with the XREADGROUP command. The
 ` + "`client_id`" + ` should be unique for each consumer of a group.`,
 		Description: `
-The field ` + "`limit`" + ` specifies the maximum number of records to be
-received per request. When more than one record is returned they are batched and
-can be split into individual messages with the ` + "`split`" + ` processor.
-
 Redis stream entries are key/value pairs, as such it is necessary to specify the
 key that contains the body of the message. All other keys/value pairs are saved
 as metadata fields.`,
@@ -32,7 +28,7 @@ as metadata fields.`,
 			docs.FieldCommon("url", "The URL of a Redis server to connect to."),
 			docs.FieldCommon("body_key", "The field key to extract the raw message from. All other keys will be stored in the message as metadata."),
 			docs.FieldCommon("streams", "A list of streams to consume from."),
-			docs.FieldCommon("limit", "The maximum number of messages to consume from a single request. When multiple messages are consumed they are processed as a [batch](/docs/configuration/batching)."),
+			docs.FieldCommon("limit", "The maximum number of messages to consume from a single request."),
 			docs.FieldCommon("client_id", "An identifier for the client connection."),
 			docs.FieldCommon("consumer_group", "An identifier for the consumer group of the stream."),
 			docs.FieldAdvanced("start_from_oldest", "If an offset is not found for a stream, determines whether to consume from the oldest available offset, otherwise messages are consumed from the latest offset."),